ABG user API/requirements.txt
flask-pymongo==2.3.0
Flask==2.3.2
Werkzeug==2.3.4
flask-mail==0.9.1
flask==2.3.2
Werkzeug==2.3.4
environ==1.0
jsonref==1.1.0
pyjwt==2.7.0
flask_bcrypt==1.0.1
python-dotenv==1.0.0
jsonschema==4.17.3
pytest==7.3.1
Six==1.16.0
flask-cors==3.0.10
openpyxl==3.1.2
requests==2.31.0
boto3==1.26.161